unsigned 2

+++ UNIX/blur_device.cu
@@ -22,13 +22,13 @@ void cudaBlurKernel(const float *raw_data, const float *blur_v, float *out_data,
     
     unsigned int thread_index = blockIdx.x * blockDim.x + threadIdx.x;
     
-    while (thread_index < (unsigned int)n_frames) {
+    while (thread_index < n_frames) {
         if (thread_index < (unsigned int)blur_v_size) {
-            for (unsigned int i = 0; i <= thread_index; ++i) {
+            for (int i = 0; i <= thread_index; ++i) {
                 out_data[thread_index] += raw_data[thread_index - i] * blur_v[i];
             }
         } else {
-            for (unsigned int i = 0; i < (unsigned int)blur_v_size; ++i) {
+            for (int i = 0; i < (unsigned int)blur_v_size; ++i) {
                 out_data[thread_index] += raw_data[thread_index - i] * blur_v[i];
             }
         }
